Store procedure testing with a cursor
Procedure that tests the student_grade cursor by fetching data from the student_grade table on a row-by-row basis. If successful, it will return the appropriate data. Otherwise, it will throw an error message with the locator_text that will help detect an error in a certain part of the program.