Stored procedure with a cursor with parameters
A stored procedure that contains 3 cursors: 1) all_students cursor with no parameters, but with appropriate columns selected from the students table. 2) one_student cursor that takes a std_id (student id) as a parameter and filters the data by the student_id in the WHERE clause. 3) one_student_course cursor that takes in 2 parameters such as student_id and student_coursename and filters these two columns in the WHERE clause. It will fetch data and return data from the cursor. If data cannot be fetched, it will return an error, using a locator_text variable to locate the section where the error originates from.